+++ ChangeLog.md
@@ -3,3 +3,4 @@
 ## 0.1
 
 * Implement Shamir secret sharing using Lagrange interpolation
+* Implement Packed variant of Shamir's scheme to share multiple secrets